1. 취득세 계산 프로세스 개선(취득세 계산결과 List 포맷으로 변경) 2. 모든 Float 자료형을 Double로 변경 - 김웅태

+++ src/main/java/com/xmonster/howtaxing/constant/CommonConstant.java
@@ -66,5 +66,5 @@ public final class CommonConstant {
     public final static String ADJUSTMENT_TARGET_AREA4 = "송파구";
 
     /* 전용면적(m2) */
-    public final static float AREA_85 = 85;
+    public final static double AREA_85 = 85;
 }
